+++ app/src/main/java/cn/wthee/pcrtool/data/enums/EnumType.kt
@@ -131,3 +131,21 @@ enum class AttrValueType {
     DOUBLE,
     PERCENT
 }
+
+
+/**
+ * 首页
+ */
+enum class OverviewType(val id: Int) {
+    CHARACTER(0),
+    EQUIP(1),
+    TOOL(2),
+    NEWS(3),
+    IN_PROGRESS_EVENT(4),
+    COMING_SOON_EVENT(5);
+
+    companion object {
+        fun getByValue(value: Int) = values()
+            .find { it.id == value } ?: CHARACTER
+    }
+}
