+++ app/src/main/java/cn/wthee/pcrtool/ui/home/Overview.kt
@@ -25,29 +25,24 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.core.content.edit
 import androidx.hilt.navigation.compose.hiltViewModel
 import cn.wthee.pcrtool.R
-import cn.wthee.pcrtool.data.db.entity.NewsTable
-import cn.wthee.pcrtool.data.db.view.*
-import cn.wthee.pcrtool.data.enums.EventType
 import cn.wthee.pcrtool.data.enums.MainIconType
+import cn.wthee.pcrtool.data.enums.OverviewType
 import cn.wthee.pcrtool.database.DatabaseUpdater
 import cn.wthee.pcrtool.ui.MainActivity
 import cn.wthee.pcrtool.ui.MainActivity.Companion.animOnFlag
 import cn.wthee.pcrtool.ui.MainActivity.Companion.navViewModel
 import cn.wthee.pcrtool.ui.NavActions
 import cn.wthee.pcrtool.ui.common.*
+import cn.wthee.pcrtool.ui.home.module.*
 import cn.wthee.pcrtool.ui.mainSP
 import cn.wthee.pcrtool.ui.theme.Dimen
-import cn.wthee.pcrtool.ui.theme.ExpandAnimation
 import cn.wthee.pcrtool.ui.theme.colorWhite
 import cn.wthee.pcrtool.ui.theme.defaultSpring
-import cn.wthee.pcrtool.ui.tool.*
-import cn.wthee.pcrtool.utils.*
+import cn.wthee.pcrtool.utils.Constants
+import cn.wthee.pcrtool.utils.VibrateUtil
+import cn.wthee.pcrtool.utils.intArrayList
 import cn.wthee.pcrtool.viewmodel.NoticeViewModel
 import cn.wthee.pcrtool.viewmodel.OverviewViewModel
-import com.google.accompanist.flowlayout.FlowColumn
-import com.google.accompanist.pager.ExperimentalPagerApi
-import com.google.accompanist.pager.HorizontalPager
-import com.google.accompanist.pager.rememberPagerState
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
 
@@ -56,21 +51,6 @@ val permissions = arrayOf(
     Manifest.permission.READ_CALENDAR,
     Manifest.permission.WRITE_CALENDAR,
 )
-
-//首页
-private enum class OverviewType(val id: Int) {
-    CHARACTER(0),
-    EQUIP(1),
-    TOOL(2),
-    NEWS(3),
-    IN_PROGRESS_EVENT(4),
-    COMING_SOON_EVENT(5);
-
-    companion object {
-        fun getByValue(value: Int) = values()
-            .find { it.id == value } ?: CHARACTER
-    }
-}
 
 /**
  * 首页纵览
@@ -219,557 +199,6 @@ fun Overview(
     }
 }
 
-/**
- * 功能模块
- */
-@Composable
-private fun ToolSection(
-    actions: NavActions,
-    isEditMode: Boolean,
-) {
-    val id = OverviewType.TOOL.id
-    Section(
-        id = id,
-        titleId = R.string.function,
-        iconType = MainIconType.FUNCTION,
-        isEditMode = isEditMode,
-        onClick = {
-            if (isEditMode)
-                editOverviewMenuOrder(id)
-            else
-                actions.toToolMore(false)
-        }
-    ) {
-        ToolMenu(actions = actions)
-    }
-}
-
-/**
- * 公告
- */
-@Composable
-private fun NewsSection(
-    actions: NavActions,
-    isEditMode: Boolean,
-    overviewViewModel: OverviewViewModel = hiltViewModel()
-) {
-    val id = OverviewType.NEWS.id
-    //公告列表
-    val newsList =
-        overviewViewModel.getNewsOverview().collectAsState(initial = null).value
-
-    Section(
-        id = id,
-        titleId = R.string.tool_news,
-        iconType = MainIconType.NEWS,
-        isEditMode = isEditMode,
-        onClick = {
-            if (isEditMode)
-                editOverviewMenuOrder(id)
-            else
-                actions.toNews()
-        }
-    ) {
-        Column {
-            if (newsList == null) {
-                for (i in 0 until 3) {
-                    NewsItem(
-                        news = NewsTable(),
-                        toNewsDetail = actions.toNewsDetail
-                    )
-                }
-            } else if (newsList.isNotEmpty()) {
-                newsList.forEach {
-                    NewsItem(
-                        news = it,
-                        toNewsDetail = actions.toNewsDetail
-                    )
-                }
-            } else {
-                CenterTipText(stringResource(id = R.string.no_data))
-            }
-        }
-    }
-}
-
-/**
- * 进行中活动
- */
-@Composable
-private fun InProgressEventSection(
-    confirmState: MutableState<Int>,
-    actions: NavActions,
-    isEditMode: Boolean,
-    overviewViewModel: OverviewViewModel = hiltViewModel()
-) {
-
-    //进行中掉落活动
-    val inProgressEventList =
-        overviewViewModel.getCalendarEventList(EventType.IN_PROGRESS)
-            .collectAsState(initial = arrayListOf()).value
-    //进行中剧情活动
-    val inProgressStoryEventList =
-        overviewViewModel.getStoryEventList(EventType.IN_PROGRESS)
-            .collectAsState(initial = arrayListOf()).value
-    //进行中卡池
-    val inProgressGachaList =
-        overviewViewModel.getGachaList(EventType.IN_PROGRESS)
-            .collectAsState(initial = arrayListOf()).value
-    //进行中免费十连
-    val inProgressFreeGachaList =
-        overviewViewModel.getFreeGachaList(EventType.IN_PROGRESS)
-            .collectAsState(initial = arrayListOf()).value
-    //进行中生日日程
-    val inProgressBirthdayList =
-        overviewViewModel.getBirthdayList(EventType.IN_PROGRESS)
-            .collectAsState(initial = arrayListOf()).value
-
-    CalendarEventLayout(
-        isEditMode,
-        EventType.IN_PROGRESS,
-        confirmState,
-        actions,
-        inProgressEventList,
-        inProgressStoryEventList,
-        inProgressGachaList,
-        inProgressFreeGachaList,
-        inProgressBirthdayList
-    )
-}
-
-/**
- * 活动预告
- */
-@Composable
-private fun ComingSoonEventSection(
-    confirmState: MutableState<Int>,
-    actions: NavActions,
-    isEditMode: Boolean,
-    overviewViewModel: OverviewViewModel = hiltViewModel()
-) {
-    //预告掉落活动
-    val comingSoonEventList =
-        overviewViewModel.getCalendarEventList(EventType.COMING_SOON)
-            .collectAsState(initial = arrayListOf()).value
-    //预告剧情活动
-    val comingSoonStoryEventList =
-        overviewViewModel.getStoryEventList(EventType.COMING_SOON)
-            .collectAsState(initial = arrayListOf()).value
-    //预告卡池
-    val comingSoonGachaList =
-        overviewViewModel.getGachaList(EventType.COMING_SOON)
-            .collectAsState(initial = arrayListOf()).value
-    //预告免费十连
-    val comingSoonFreeGachaList =
-        overviewViewModel.getFreeGachaList(EventType.COMING_SOON)
-            .collectAsState(initial = arrayListOf()).value
-    //生日
-    val comingSoonBirthdayList =
-        overviewViewModel.getBirthdayList(EventType.COMING_SOON)
-            .collectAsState(initial = arrayListOf()).value
-
-    CalendarEventLayout(
-        isEditMode,
-        EventType.COMING_SOON,
-        confirmState,
-        actions,
-        comingSoonEventList,
-        comingSoonStoryEventList,
-        comingSoonGachaList,
-        comingSoonFreeGachaList,
-        comingSoonBirthdayList
-    )
-}
-
-
-/**
- * 角色
- */
-@OptIn(ExperimentalPagerApi::class)
-@Composable
-private fun CharacterSection(
-    actions: NavActions,
-    isEditMode: Boolean,
-    overviewViewModel: OverviewViewModel = hiltViewModel()
-) {
-    val id = OverviewType.CHARACTER.id
-
-    //角色总数
-    val characterCount =
-        overviewViewModel.getCharacterCount().collectAsState(initial = 0).value
-    //角色列表
-    val characterList =
-        overviewViewModel.getCharacterInfoList().collectAsState(initial = arrayListOf()).value
-    Section(
-        id = id,
-        titleId = R.string.character,
-        iconType = MainIconType.CHARACTER,
-        hintText = characterCount.toString(),
-        contentVisible = characterList.isNotEmpty(),
-        isEditMode = isEditMode,
-        onClick = {
-            if (isEditMode)
-                editOverviewMenuOrder(id)
-            else
-                actions.toCharacterList()
-        }
-    ) {
-        if (characterList.isNotEmpty()) {
-            HorizontalPager(
-                count = characterList.size,
-                state = rememberPagerState(),
-                modifier = Modifier
-                    .padding(vertical = Dimen.mediumPadding)
-                    .fillMaxWidth(),
-                contentPadding = PaddingValues(horizontal = Dimen.largePadding),
-                itemSpacing = Dimen.mediumPadding
-            ) { index ->
-                val unitId = if (characterList.isEmpty()) 0 else characterList[index].id
-                MainCard(
-                    modifier = Modifier.width(getItemWidth()),
-                    onClick = {
-                        actions.toCharacterDetail(unitId)
-                    }
-//                    elevation = CardDefaults.cardElevation(0.dp),
-//                    shape = MaterialTheme.shapes.medium,
-//                    colors = CardDefaults.cardColors(containerColor = MaterialTheme.colorScheme.background)
-                ) {
-                    ImageCompose(
-                        data = ImageResourceHelper.getInstance().getMaxCardUrl(unitId),
-                        ratio = RATIO
-                    )
-                }
-            }
-        }
-    }
-}
-
-/**
- * 装备
- */
-@Composable
-private fun EquipSection(
-    actions: NavActions,
-    isEditMode: Boolean,
-    overviewViewModel: OverviewViewModel = hiltViewModel()
-) {
-    val id = OverviewType.EQUIP.id
-    val equipSpanCount =
-        ScreenUtil.getWidth() / (Dimen.iconSize + Dimen.largePadding * 2).value.dp2px
-    //装备总数
-    val equipCount = overviewViewModel.getEquipCount().collectAsState(initial = 0).value
-    //装备列表
-    val equipList = overviewViewModel.getEquipList(maxOf(1, equipSpanCount) * 2)
-        .collectAsState(initial = arrayListOf()).value
-
-    Section(
-        id = id,
-        titleId = R.string.tool_equip,
-        iconType = MainIconType.EQUIP,
-        hintText = equipCount.toString(),
-        contentVisible = equipList.isNotEmpty(),
-        isEditMode = isEditMode,
-        onClick = {
-            if (isEditMode)
-                editOverviewMenuOrder(id)
-            else
-                actions.toEquipList()
-        }
-    ) {
-        VerticalGrid(
-            spanCount = maxOf(1, equipSpanCount),
-            modifier = Modifier.padding(horizontal = Dimen.commonItemPadding)
-        ) {
-            if (equipList.isNotEmpty()) {
-                equipList.forEach {
-                    Box(
-                        modifier = Modifier
-                            .padding(Dimen.mediumPadding)
-                            .fillMaxWidth(),
-                        contentAlignment = Alignment.Center
-                    ) {
-                        IconCompose(
-                            data = ImageResourceHelper.getInstance()
-                                .getEquipPic(it.equipmentId)
-                        ) {
-                            actions.toEquipDetail(it.equipmentId)
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-/**
- * 活动
- */
-@Composable
-private fun CalendarEventLayout(
-    isEditMode: Boolean,
-    calendarType: EventType,
-    confirmState: MutableState<Int>,
-    actions: NavActions,
-    eventList: List<CalendarEvent>,
-    storyEventList: List<EventData>,
-    gachaList: List<GachaInfo>,
-    freeGachaList: List<FreeGachaInfo>,
-    birthdayList: List<BirthdayData>
-) {
-    val id = if (calendarType == EventType.IN_PROGRESS) {
-        OverviewType.IN_PROGRESS_EVENT.id
-    } else {
-        OverviewType.COMING_SOON_EVENT.id
-    }
-    val spanCount = ScreenUtil.getWidth() / getItemWidth().value.dp2px
-    val isNotEmpty =
-        eventList.isNotEmpty() || storyEventList.isNotEmpty() || gachaList.isNotEmpty() || freeGachaList.isNotEmpty() || birthdayList.isNotEmpty()
-    val titleId = if (calendarType == EventType.IN_PROGRESS) {
-        R.string.tool_calendar_inprogress
-    } else {
-        R.string.tool_calendar_comming
-    }
-
-
-    if (isEditMode || isNotEmpty) {
-        Section(
-            id = id,
-            titleId = titleId,
-            iconType = if (calendarType == EventType.IN_PROGRESS) MainIconType.CALENDAR_TODAY else MainIconType.CALENDAR,
-            rightIconType = if (confirmState.value == calendarType.type) MainIconType.CLOSE else MainIconType.MAIN,
-            isEditMode = isEditMode,
-            onClick = {
-                if (isEditMode) {
-                    editOverviewMenuOrder(id)
-                } else {
-                    //弹窗确认
-                    if (confirmState.value == calendarType.type) {
-                        confirmState.value = 0
-                    } else {
-                        confirmState.value = calendarType.type
-                    }
-                }
-            }
-        ) {
-            ExpandAnimation(visible = confirmState.value == calendarType.type) {
-                CalendarEventOperation(
-                    confirmState,
-                    eventList,
-                    storyEventList,
-                    gachaList,
-                    freeGachaList,
-                    birthdayList
-                )
-            }
-            VerticalGrid(
-                spanCount = spanCount,
-                modifier = Modifier.padding(top = Dimen.mediumPadding)
-            ) {
-                gachaList.forEach {
-                    GachaItem(it, actions.toCharacterDetail)
-                }
-                storyEventList.forEach {
-                    StoryEventItem(it, actions.toCharacterDetail, actions.toAllPics)
-                }
-                eventList.forEach {
-                    CalendarEventItem(it)
-                }
-                freeGachaList.forEach {
-                    FreeGachaItem(it)
-                }
-                birthdayList.forEach {
-                    BirthdayItem(it, actions.toCharacterDetail)
-                }
-            }
-        }
-    }
-}
-
-/**
- * 添加日历确认
- */
-@Composable
-private fun CalendarEventOperation(
-    confirmState: MutableState<Int>,
-    eventList: List<CalendarEvent>,
-    storyEventList: List<EventData>,
-    gachaList: List<GachaInfo>,
-    freeGachaList: List<FreeGachaInfo>,
-    birthdayList: List<BirthdayData>,
-) {
-    val context = LocalContext.current
-    val regionName = getRegionName(MainActivity.regionType)
-
-    // 添加日历确认
-    MainCard(
-        modifier = Modifier.padding(
-            horizontal = Dimen.largePadding,
-            vertical = Dimen.mediumPadding
-        )
-    ) {
-        FlowColumn {
-            //添加日历
-            Row(
-                modifier = Modifier
-                    .padding(horizontal = Dimen.largePadding, vertical = Dimen.mediumPadding)
-                    .clip(MaterialTheme.shapes.medium)
-                    .clickable {
-                        VibrateUtil(context).single()
-                        checkPermissions(context, permissions, false) {
-                            val allEvents = arrayListOf<SystemCalendarEventData>()
-                            //掉落活动
-                            eventList.forEach {
-                                allEvents.add(
-                                    SystemCalendarEventData(
-                                        it.startTime,
-                                        it.endTime,
-                                        getTypeDataToString(it)
-                                    )
-                                )
-                            }
-                            //剧情活动
-                            storyEventList.forEach {
-                                allEvents.add(
-                                    SystemCalendarEventData(
-                                        it.startTime,
-                                        it.endTime,
-                                        it.getEventTitle()
-                                    )
-                                )
-                            }
-                            //卡池
-                            gachaList.forEach {
-                                allEvents.add(
-                                    SystemCalendarEventData(
-                                        it.startTime,
-                                        it.endTime,
-                                        it.getDesc()
-                                    )
-                                )
-                            }
-                            //免费十连
-                            freeGachaList.forEach {
-                                allEvents.add(
-                                    SystemCalendarEventData(
-                                        it.startTime,
-                                        it.endTime,
-                                        it.getDesc()
-                                    )
-                                )
-                            }
-                            //免费十连
-                            freeGachaList.forEach {
-                                allEvents.add(
-                                    SystemCalendarEventData(
-                                        it.startTime,
-                                        it.endTime,
-                                        it.getDesc()
-                                    )
-                                )
-                            }
-                            //生日日程
-                            birthdayList.forEach {
-                                allEvents.add(
-                                    SystemCalendarEventData(
-                                        it.getStartTime(),
-                                        it.getEndTime(),
-                                        it.getDesc()
-                                    )
-                                )
-                            }
-                            //添加至系统日历
-                            SystemCalendarHelper().insertEvents(allEvents)
-
-                            confirmState.value = 0
-                        }
-                    }
-                    .padding(Dimen.mediumPadding),
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                IconCompose(data = MainIconType.ADD_CALENDAR, size = Dimen.fabIconSize)
-                MainText(
-                    text = stringResource(R.string.add_to_calendar),
-                    modifier = Modifier.padding(horizontal = Dimen.smallPadding)
-                )
-            }
-
-            //复制至剪贴板
-            Row(
-                modifier = Modifier
-                    .padding(horizontal = Dimen.largePadding, vertical = Dimen.mediumPadding)
-                    .clip(MaterialTheme.shapes.medium)
-                    .clickable {
-                        VibrateUtil(context).single()
-                        var allText = ""
-                        //掉落活动
-                        var eventText = ""
-                        eventList.forEach {
-                            val date = getCalendarEventDateText(it.startTime, it.endTime)
-                            eventText += "• $date\n${getTypeDataToString(it)}\n"
-                        }
-                        if (eventText != "") {
-                            allText += "▶ 掉落活动\n$eventText\n"
-                        }
-
-                        //剧情活动
-                        var storyText = ""
-                        storyEventList.forEach {
-                            val date = getCalendarEventDateText(it.startTime, it.endTime)
-                            storyText += "• $date\n${it.getEventTitle()}"
-                        }
-                        if (storyText != "") {
-                            allText += "▶ 剧情活动\n$storyText\n\n"
-                        }
-
-                        //卡池
-                        var gachaText = ""
-                        gachaList.forEach {
-                            val date = getCalendarEventDateText(it.startTime, it.endTime)
-                            gachaText += "• $date\n${it.getDesc()}"
-
-                        }
-                        if (gachaText != "") {
-                            allText += "▶ 卡池信息\n$gachaText\n\n"
-                        }
-
-                        //免费十连
-                        var freeGachaText = ""
-                        freeGachaList.forEach {
-                            val date = getCalendarEventDateText(it.startTime, it.endTime)
-                            freeGachaText += "• $date\n${it.getDesc()}"
-
-                        }
-                        if (freeGachaText != "") {
-                            allText += "▶ 免费十连\n$freeGachaText\n\n"
-                        }
-
-                        //免费十连
-                        var birthdayText = ""
-                        birthdayList.forEach {
-                            val date = getCalendarEventDateText(it.getStartTime(), it.getEndTime())
-                            birthdayText += "• $date\n${it.getDesc()}"
-
-                        }
-                        if (birthdayText != "") {
-                            allText += "▶ 角色生日\n$birthdayText\n"
-                        }
-                        //复制
-                        copyText(context, "——$regionName——\n\n$allText")
-                        confirmState.value = 0
-                    }
-                    .padding(Dimen.mediumPadding),
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                IconCompose(data = MainIconType.COPY, size = Dimen.fabIconSize)
-                MainText(
-                    text = stringResource(R.string.copy_event),
-                    modifier = Modifier.padding(horizontal = Dimen.smallPadding)
-                )
-            }
-        }
-
-    }
-}
 
 /**
  * 数据切换选择弹窗
@@ -871,11 +300,12 @@ private fun ChangeDbCompose(
     }
 }
 
+
 /**
  * 标题、内容
  */
 @Composable
-private fun Section(
+fun Section(
     id: Int,
     @StringRes titleId: Int,
     iconType: MainIconType,
@@ -986,53 +416,11 @@ private fun Section(
 
 }
 
+
 /**
- * 获取事项信息
+ * 编辑排序
  */
-private fun getTypeDataToString(data: CalendarEvent): String {
-    var eventTitle = ""
-    when (data.type) {
-        "1" -> {
-            //露娜塔
-            eventTitle = "露娜塔"
-        }
-        "-1" -> {
-            //特殊地下城
-            eventTitle = "特殊地下城"
-        }
-        else -> {
-            //正常活动
-            val list = data.type.intArrayList
-            list.forEachIndexed { index, type ->
-                val title = when (type) {
-                    31, 41 -> "普通关卡"
-                    32, 42 -> "困难关卡"
-                    39, 49 -> "高难关卡"
-                    34 -> "探索"
-                    37 -> "圣迹调查"
-                    38 -> "神殿调查"
-                    45 -> "地下城"
-                    else -> ""
-                }
-                val multiple = data.getFixedValue()
-                eventTitle += title + (if (type > 40) "玛那掉落量" else "掉落量") + (if ((multiple * 10).toInt() % 10 == 0) {
-                    multiple.toInt().toString()
-                } else {
-                    multiple.toString()
-                }) + "倍"
-                if (index != list.size - 1) {
-                    eventTitle += "\n"
-                }
-            }
-        }
-    }
-
-    return eventTitle
-}
-
-
-//编辑排序
-private fun editOverviewMenuOrder(id: Int) {
+fun editOverviewMenuOrder(id: Int) {
     val sp = mainSP()
     val orderStr = sp.getString(Constants.SP_OVERVIEW_ORDER, "") ?: ""
     val idStr = "$id-"
@@ -1050,14 +438,3 @@ private fun editOverviewMenuOrder(id: Int) {
         navViewModel.overviewOrderData.postValue(edited)
     }
 }
-
-/**
- * 日历日程时间范围文本
- */
-private fun getCalendarEventDateText(
-    startTime: String,
-    endTime: String
-) = startTime.formatTime.fixJpTime.substring(
-    0,
-    10
-) + " ~ " + endTime.formatTime.fixJpTime.substring(0, 10)
